requirements.txt
pyserial
pynmea2
numpy
matplotlib
